+++ pkg/cl/file-author-unix.go
@@ -6,14 +6,20 @@ import (
 	"os/user"
 	"strconv"
 	"syscall"
+
+	"github.com/ohler55/slip"
 )
 
-func getFileOwner(path string) (string, error) {
-	var stat syscall.Stat_t
-	if err := syscall.Stat(path, &stat); err == nil {
-		if u, err := user.LookupId(strconv.Itoa(int(stat.Uid))); err == nil {
-			return u.Username, nil
+func getFileOwner(s *slip.Scope, path string) string {
+	var (
+		stat syscall.Stat_t
+		err  error
+	)
+	if err = syscall.Stat(path, &stat); err == nil {
+		var u *user.User
+		if u, err = user.LookupId(strconv.Itoa(int(stat.Uid))); err == nil {
+			return u.Username
 		}
 	}
-	return "", nil
+	panic(slip.FileErrorNew(s, 0, slip.String(path), "file-author failed. %s", err))
 }