+++ pkg/cl/file-author.go
@@ -42,13 +42,8 @@ func (f *FileAuthor) Call(s *slip.Scope, args slip.List, depth int) (result slip
 	if !ok {
 		slip.TypePanic(s, depth, "filepath", args[0], "string")
 	}
-
-	owner, err := getFileOwner(string(path))
-	if err != nil {
-		slip.FilePanic(s, depth, path, "%s", err)
+	if owner := getFileOwner(s, string(path)); owner != "" {
+		result = slip.String(owner)
 	}
-	if owner != "" {
-		return slip.String(owner)
-	}
-	return nil
+	return
 }